requirements.txt
# AI Overview Monitor — Python dependencies
# Install with: pip install -r requirements.txt

# Core
pyyaml>=6.0

# Browser engines (install the one you need)
# OpenClaw: install via npm (npm install -g openclaw@latest)
# Playwright:
playwright>=1.40.0
# Selenium (universal fallback):
selenium>=4.15.0
webdriver-manager>=4.0.0

# For HTML parsing (used in citation extraction)
beautifulsoup4>=4.12.0
lxml>=5.0.0
